feat: add the Grand Unified PR
This PR does the following:

- adds a code formatter (Black)
- adds a linter (ruff)
- adds tests (Pytest)
- documents code by adding Python docstring comments to functions and
  methods
- attempts to address https://github.com/numberscope/backscope/issues/69
- attempts to address some of the feedback in
  https://github.com/numberscope/backscope/pull/74

There are some minor changes that aren't worth delving into in the
commit message. The API has changed somewhat, so if we were using
Semantic Versioning, this would be a breaking change / major version
upgrade.

I will go into more detail in the pull request description on GitHub.