fix: Update format for search by ID (#140)
This is a single-commit version of #135. It addresses #133 by changing the search query in `get_oeis_name_and_values`. The Requests library automatically URL-encodes the query, so the `:` in the query gets sent as `%3A`.

This pull request is based on #127, which should be merged first. The history associated with this pull request starts at commit 1eec7cd.