Stop using the first B-file comment as stopgap name
The `get_oeis_name_and_values` endpoint only looks up a sequence's name
if the name field in the database is either empty or set to the
placeholder name. As a result, this endpoint can't overwrite the stopgap
name, even though we want it to.

It seems like this problem should've existed before we started
developing the local metadata feature, so it would be useful to know why
we didn't notice it before. This is hard to investigate, though, because
the OEIS search format change prevents the old code from working as-is.