Store sequence metadata locally (#164)
* Get name and references from local metadata

Use local sequence file reads as drop-in replacements for OEIS searches.
Update tests to use sequences' real names instead of placeholders,
because names and raw references are now written to the database almost
immediately after `fetch_metadata` is called. Since we're still fetching
names and raw references asynchronously, this could make the tests
flaky.

* Add instructions for downloading the OEIS metadata

* Stop using the first B-file comment as stopgap name

The `get_oeis_name_and_values` endpoint only looks up a sequence's name
if the name field in the database is either empty or set to the
placeholder name. As a result, this endpoint can't overwrite the stopgap
name, even though we want it to.

It seems like this problem should've existed before we started
developing the local metadata feature, so it would be useful to know why
we didn't notice it before. This is hard to investigate, though, because
the OEIS search format change prevents the old code from working as-is.

* Strip sequence name in `get_oeis_name_and_values`

Also, add a test of that would've caught this bug.

* Add `ripgrep` as a dependency

* fix: Restore search_oeis endpoint, at least somewhat

---------

Co-authored-by: Aaron Fenyes <aaron.fenyes@fareycircles.ooo>
Co-authored-by: Glen Whitney <glen@studioinfinity.org>